requirements.txt
asgiref==3.7.2
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2023.4
